# docs/apispec_1.json
{
...
  "info": {
    "description": "A RESTful API for researchers to track arXiv papers, manage research notes, and use spaced repetition learning.",
    "termsOfService": "",
    "title": "ScholarTrack API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/notes": {
      "get": {
        "parameters": [
          {
            "description": "Filter by arXiv paper ID",
            "example": "2301.07041",
            "in": "query",
            "name": "paper_id",
            "required": false,
            "type": "string"
          },
          {
            "default": 1,
            "description": "Page number",
            "in": "query",
            "name": "page",
            "required": false,
            "type": "integer"
          },
          {
            "default": 20,
            "description": "Results per page (max 100)",
            "in": "query",
            "name": "per_page",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of notes"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "List all notes for the authenticated user.",
        "tags": [
          "Notes"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "content": {
                  "example": "The paper demonstrates that multi-head attention can be pruned without significant performance loss.",
                  "type": "string"
                },
                "paper_id": {
                  "description": "arXiv ID to link the note to a paper",
                  "example": "2301.07041",
                  "type": "string"
                },
                "title": {
                  "example": "Key findings on attention mechanisms",
                  "type": "string"
                }
              },
              "required": [
                "title",
                "content"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Note created successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "404": {
            "description": "Linked paper not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Create a new research note.",
        "tags": [
          "Notes"
        ]
      }
    },
    "/api/notes/{note_id}": {
      "delete": {
        "parameters": [
          {
            "description": "The note ID",
            "example": 1,
            "in": "path",
            "name": "note_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Note deleted"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Note not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Delete a note.",
        "tags": [
          "Notes"
        ]
      },
      "get": {
        "parameters": [
          {
            "description": "The note ID",
            "example": 1,
            "in": "path",
            "name": "note_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Note details"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Note not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get a specific note by ID.",
        "tags": [
          "Notes"
        ]
      },
      "put": {
        "parameters": [
          {
            "description": "The note ID",
            "example": 1,
            "in": "path",
            "name": "note_id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "content": {
                  "example": "Revised findings after reading the supplementary material.",
                  "type": "string"
                },
                "title": {
                  "example": "Updated title on attention mechanisms",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note updated successfully"
          },
          "400": {
            "description": "Request body is required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Note not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Update an existing note.",
        "tags": [
          "Notes"
        ]
      }
    },
...
    "/api/papers/search": {
      "get": {
        "parameters": [
          {
            "description": "Search query string",
            "example": "transformer attention mechanism",
            "in": "query",
            "name": "q",
            "required": true,
            "type": "string"
          },
          {
            "description": "arXiv category filter",
            "example": "cs.AI",
            "in": "query",
            "name": "category",
            "required": false,
            "type": "string"
          },
          {
            "description": "Start date filter (YYYY-MM-DD)",
            "example": "2025-01-01",
            "in": "query",
            "name": "date_from",
            "required": false,
            "type": "string"
          },
          {
            "description": "End date filter (YYYY-MM-DD)",
            "example": "2025-12-31",
            "in": "query",
            "name": "date_to",
            "required": false,
            "type": "string"
          },
          {
            "default": 1,
            "description": "Page number",
            "in": "query",
            "name": "page",
            "required": false,
            "type": "integer"
          },
          {
            "default": 20,
            "description": "Results per page (max 100)",
            "in": "query",
            "name": "per_page",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching papers with pagination"
          },
          "400": {
            "description": "Missing required query parameter \"q\""
          }
        },
        "summary": "Search for papers on arXiv.",
        "tags": [
          "Paper Discovery"
        ]
      }
    },
...
    "/api/papers/{arxiv_id}": {
      "get": {
        "parameters": [
          {
            "description": "The arXiv paper identifier",
            "example": "2301.07041",
            "in": "path",
            "name": "arxiv_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Paper details (includes is_saved field if authenticated)"
          },
          "404": {
            "description": "Paper not found"
          }
        },
        "summary": "Get details of a specific paper by arXiv ID.",
        "tags": [
          "Paper Discovery"
        ]
      }
    },
    "/api/papers/{arxiv_id}/notes": {
      "get": {
        "parameters": [
          {
            "description": "The arXiv paper identifier",
            "example": "2301.07041",
            "in": "path",
            "name": "arxiv_id",
            "required": true,
            "type": "string"
          },
          {
            "default": 1,
            "description": "Page number",
            "in": "query",
            "name": "page",
            "required": false,
            "type": "integer"
          },
          {
            "default": 20,
            "description": "Results per page (max 100)",
            "in": "query",
            "name": "per_page",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of notes for the paper"
          },
          "404": {
            "description": "Paper not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "List all notes for a specific paper.",
        "tags": [
          "Notes"
        ]
      }
    },
    "/api/papers/{arxiv_id}/save": {
      "delete": {
        "parameters": [
          {
            "description": "The arXiv paper identifier",
            "example": "2301.07041",
            "in": "path",
            "name": "arxiv_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Paper removed from library"
          },
          "404": {
            "description": "Paper not found in library"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Remove a paper from the user's library.",
        "tags": [
          "Paper Management"
        ]
      },
      "post": {
        "parameters": [
          {
            "description": "The arXiv paper identifier",
            "example": "2301.07041",
            "in": "path",
            "name": "arxiv_id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "properties": {
                "memo": {
                  "example": "Interesting approach to multi-head attention",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Paper saved successfully"
          },
          "404": {
            "description": "Paper not found"
          },
          "409": {
            "description": "Paper already saved"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Save a paper to the user's library.",
        "tags": [
          "Paper Management"
        ]
      }
    },
...
    "/api/search/all": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "n_results": {
                  "default": 10,
                  "description": "Number of results per source (max 50)",
                  "type": "integer"
                },
                "query": {
                  "example": "attention mechanism improvements",
                  "type": "string"
                }
              },
              "required": [
                "query"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Combined results from papers and notes sorted by relevance",
            "schema": {
              "properties": {
                "data": {
                  "items": {
                    "properties": {
                      "relevance_score": {
                        "type": "number"
                      },
                      "type": {
                        "enum": [
                          "paper",
                          "note"
                        ],
                        "type": "string"
                      }
                    },
                    "type": "object"
                  },
                  "type": "array"
                },
                "query": {
                  "type": "string"
                },
                "source": {
                  "example": "all",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Missing or empty query field"
          },
          "401": {
            "description": "Missing or invalid JWT token"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Search across both papers and notes, merged by relevance.",
        "tags": [
          "Semantic Search"
        ]
      }
    },
    "/api/search/notes": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "n_results": {
                  "default": 5,
                  "description": "Number of results to return (max 50)",
                  "type": "integer"
                },
                "query": {
                  "example": "data augmentation techniques",
                  "type": "string"
                }
              },
              "required": [
                "query"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Semantic search results from user notes",
            "schema": {
              "properties": {
                "data": {
                  "items": {
                    "properties": {
                      "content": {
                        "type": "string"
                      },
                      "id": {
                        "type": "integer"
                      },
                      "relevance_score": {
                        "type": "number"
                      },
                      "title": {
                        "type": "string"
                      },
                      "type": {
                        "example": "note",
                        "type": "string"
                      }
                    },
                    "type": "object"
                  },
                  "type": "array"
                },
                "query": {
                  "type": "string"
                },
                "source": {
                  "example": "user_notes",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Missing or empty query field"
          },
          "401": {
            "description": "Missing or invalid JWT token"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Semantic search over the current user's notes.",
        "tags": [
          "Semantic Search"
        ]
      }
    },
    "/api/search/papers": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "n_results": {
                  "default": 10,
                  "description": "Number of results to return (max 50)",
                  "type": "integer"
                },
                "query": {
                  "example": "transformer attention mechanism",
                  "type": "string"
                }
              },
              "required": [
                "query"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Semantic search results from paper abstracts",
            "schema": {
              "properties": {
                "data": {
                  "items": {
                    "properties": {
                      "abstract": {
                        "type": "string"
                      },
                      "arxiv_id": {
                        "type": "string"
                      },
                      "relevance_score": {
                        "description": "Similarity score 0-1 (higher = more relevant)",
                        "type": "number"
                      },
                      "title": {
                        "type": "string"
                      },
                      "type": {
                        "example": "paper",
                        "type": "string"
                      }
                    },
                    "type": "object"
                  },
                  "type": "array"
                },
                "query": {
                  "type": "string"
                },
                "source": {
                  "example": "paper_abstracts",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Missing or empty query field"
          }
        },
        "summary": "Semantic search over paper abstracts using ChromaDB.",
        "tags": [
          "Semantic Search"
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "JWT Bearer token. Format: \"Bearer {token}\"",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}